added draw,discard,exhaust,hand pile enum to card, and adding discarded cards back into the hand after a turn

+++ include/main_ui.h
@@ -39,7 +39,7 @@ public:
     void update_labels();
     void update_card_display();
     void _on_end_turn_pressed();
-    void _on_card_pressed(size_t index);
+    void _on_card_pressed(int index);
 };
 
 } // namespace godot
